# services/backend/docs/swagger.yaml
basePath: /
definitions:
  handler.ChannelRef:
    properties:
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
...
  handler.NewsGroup:
    properties:
      duplicate_count:
        type: integer
      duplicates:
        items:
          $ref: '#/definitions/handler.NewsItem'
        type: array
      main:
        $ref: '#/definitions/handler.NewsItem'
    type: object
  handler.NewsItem:
    properties:
      body:
        type: string
      channel:
        $ref: '#/definitions/handler.ChannelRef'
      created_at:
        type: string
      id:
        type: integer
      message_link:
        type: string
      topics:
        items:
          $ref: '#/definitions/handler.TopicRef'
        type: array
    type: object
...
  handler.TopicRef:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Channel:
    properties:
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.News:
    properties:
      body:
        type: string
      channel:
        $ref: '#/definitions/models.Channel'
      channelID:
        type: integer
      createdAt:
        type: string
      duplicates:
        items:
          $ref: '#/definitions/models.News'
        type: array
      id:
        type: integer
      messageLink:
        type: string
      stopThemes:
        items:
          $ref: '#/definitions/models.StopTheme'
        type: array
      stopThemesCheckedAt:
        type: string
      topics:
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      topicsCheckedAt:
        type: string
      withoutTopics:
        type: boolean
    type: object
  models.StopTheme:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
    type: object
  models.Topic:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
    type: object
  models.User:
    properties:
      channels:
        items:
          $ref: '#/definitions/models.Channel'
        type: array
      id:
        type: integer
      stopThemes:
        items:
          $ref: '#/definitions/models.StopTheme'
        type: array
      topics:
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for SmartFlow application
  title: SmartFlow REST API
  version: "1.0"
paths:
...
  /users/{id}/blocked-news/{stopThemeId}:
    get:
      description: Returns news from user's channels blocked by a specific stop-theme,
        grouped by duplicates.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Stop Theme ID
        in: path
        name: stopThemeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.NewsGroup'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user's blocked news
      tags:
      - user-news
  /users/{id}/channels:
    get:
      description: Returns all channels the user is subscribed to
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Channel'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user channels
      tags:
      - user-channels
    post:
      consumes:
      - application/json
      description: Subscribes the user to a channel. Creates the channel if it doesn't
        exist.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel data (Link and Name)
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add channel to user
      tags:
      - user-channels
  /users/{id}/channels/{channelId}:
    delete:
      description: Unsubscribes the user from a channel. Deletes the channel if no
        other users are subscribed.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove channel from user
      tags:
      - user-channels
  /users/{id}/news:
    get:
      description: Returns news feed grouped by duplicates. Excludes stop-themed news.
        Optional filter by topic_id.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by topic ID
        in: query
        name: topic_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.NewsGroup'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user news feed
      tags:
      - user-news
...
  /users/{id}/topics:
    get:
      description: Returns all topics (categories) the user is subscribed to
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Topic'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user topics
      tags:
      - user-topics
    post:
      consumes:
      - application/json
      description: Subscribes the user to a topic. Creates the topic if it doesn't
        exist.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Topic data (Name)
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Topic'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add topic to user
      tags:
      - user-topics
  /users/{id}/topics/{topicId}:
    delete:
      description: Unsubscribes the user from a topic. Deletes the topic if no other
        users are subscribed.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Topic ID
        in: path
        name: topicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove topic from user
      tags:
      - user-topics
swagger: "2.0"